requirements-txt.txt
# Core dependencies
numpy>=1.19.0
opencv-python>=4.5.0
PyYAML>=5.4.0
pandas>=1.2.0
matplotlib>=3.3.0
scipy>=1.6.0

# YOLO detection
ultralytics>=8.0.0
torch>=1.10.0
torchvision>=0.11.0

# Tracking
filterpy>=1.4.5

# GUI
PyQt5>=5.15.0
